+++ App.tsx
@@ -15,23 +15,11 @@ import {
   getAuth
 } from 'firebase/auth';
 
-let app, auth;
 
-if (!getApps().length) {
-  try {
-    app = initializeApp(firebaseConfig);
-    auth = initializeAuth(app, {
-      persistence: getReactNativePersistence(ReactNativeAsyncStorage)
-    });
-
-  } catch (error) {
-    console.log('Error initializing app: ' + error);
-  }
-} else {
-  app = getApp();
-  auth = getAuth(app);
-}
-export const authService = getAuth(app);
+export const appService = initializeApp(firebaseConfig)
+const auth = initializeAuth(appService, {
+  persistence: getReactNativePersistence(ReactNativeAsyncStorage)
+});
 
 
 export default function App() {