+++ src/components/AuthenticationError/styles.ts
@@ -1,0 +1,7 @@
+import styled from "styled-components/native";
+
+export const BoxErrorText = styled.View`
+    align-items: center;
+    width: 100%;
+    margin-top: 4%;
+`