+++ src/components/AuthenticationMethods/styles.ts
@@ -1,35 +1,7 @@
 import styled from "styled-components/native"
-import { theme } from "../../../../styles/theme"
+import { theme } from "../../styles/theme"
 
-export const ContainerItems = styled.View`
-   align-items: center;
-   width: 100%;
-    
-   `
-export const Inputbox = styled.View`
-    width: 90%;
-    align-items: center;
-    justify-content: center;
-    margin-bottom: 4%;
-`
-export const FourgoutPassBox = styled.View`
-    align-items: flex-end;
-    justify-items: right;
-    width: 90%;
-    margin-bottom: 3%;
-`
-export const ForgoutPass= styled.TouchableOpacity`
-       background-color: none;
-   `
-export const ButtonBox = styled.View`
-    width: 80%;
-    height: 45px;
-    justify-content: center;
-    align-items: center;
-   margin-bottom: 4%;
-   border-radius: 6px;
-`
-export const ContainerGoogle = styled.View`
+export const AuthenticationContainer = styled.View`
     width: 90%;
     justify-content: center;
     align-items: center;
@@ -58,6 +30,7 @@ export const BoxGoogle = styled.View`
 `
 export const BoxText = styled.View`
     width: 100%;
+    margin-top: 8%;
     flex-direction: row;
     justify-content: center;
 `
@@ -67,9 +40,4 @@ export const TextItem = styled.View`
 export const ButtonItem = styled.View`
     width:28%;
     height: 24px;
-`
-export const BoxErrorText = styled.View`
-    align-items: center;
-    width: 100%;
-    margin-top: 4%;
 `