+++ src/screens/Home/components/FreeStorage/styles.ts
@@ -1,0 +1,11 @@
+import styled from "styled-components/native"
+
+export const FreeStorageContainer = styled.View`
+  width: 50%;
+  align-items: center;
+  flex-direction: column;
+  margin-bottom: 8%;
+`
+export const FreeStorageBox = styled.View`
+  margin-bottom: 2%;
+`