+++ src/screens/Home/components/SearchComponents/styles.ts
@@ -1,0 +1,32 @@
+import styled from "styled-components/native"
+
+
+export const InputContainer = styled.View`
+  width: 100%;
+  align-items: center;
+`
+export const SearchBox = styled.View`
+flex-direction: row;
+`
+export const SearchItem = styled.View`
+ width: 70%;
+ margin: 0 2% 0 4%;
+`
+export const ButtonSearch = styled.View`
+  width: 20%;
+  height: 45px;
+`
+export const ContainerItem = styled.View`
+  align-items: center;
+  width: 100%;
+`
+export const RenderItem = styled.ScrollView`
+  width: 100%;
+  max-height: 300px;
+  margin-top: 8%;
+`
+export const ButtonBox = styled.View`
+  width: 80%;
+  height: 36px;
+  margin-bottom: 4%;
+`