+++ src/screens/Home/components/UploadButton/styles.ts
@@ -1,0 +1,7 @@
+import styled from "styled-components/native";
+
+export const ButtonBox = styled.View`
+  width: 160px;
+  height: 160px;
+  margin-bottom: 8%;
+`