+++ src/screens/Home/index.tsx
@@ -1,29 +1,28 @@
 import { Text } from "../../components/Text/Text";
-import { Octicons } from "@expo/vector-icons";
 import { StackPramsList } from "../../routes/AppRoutes";
 import { RouteProp } from "@react-navigation/native";
 import {
-  ButtonBox,
-  ButtonContainer,
+  ContainerItems,
   ButtonSearch,
   Container,
-  FreeStorageBox,
-  FreeStorageContainer,
   InputContainer,
   SearchItem,
   TextContainer,
   TextWelcome,
 } from "./styles";
-import ButtonSelect from "../../components/UI/ButtonSelect/ButtonSelect";
 import InputItem from "../../components/UI/Input/Input";
 import Button from "../../components/UI/Button/Button";
+import UploadButton from "./components/UploadButton";
+import FreeStorage from "./components/FreeStorage";
+import SearchComponent from "./components/SearchComponents";
 
 type HomeRouteProp = RouteProp<StackPramsList, "Home">;
 interface Props {
   route: HomeRouteProp;
+  id: number
 }
 
-export default function Home({ route }: Props) {
+export default function Home({ route, id }: Props) {
   const userEmail = route.params?.userEmail;
   return (
     <Container>
@@ -39,36 +38,11 @@ export default function Home({ route }: Props) {
           </Text>
         </TextWelcome>
       </TextContainer>
-      <ButtonContainer>
-        <ButtonBox>
-          <ButtonSelect buttoncolor="primary" onPress={console.log("teste")}>
-            <Octicons name="upload" size={72} color="white" />
-          </ButtonSelect>
-        </ButtonBox>
-        <FreeStorageContainer>
-          <FreeStorageBox>
-            <Text>Free Storage</Text>
-          </FreeStorageBox>
-          <FreeStorageBox>
-            <Text>{"size"}</Text>
-          </FreeStorageBox>
-          <FreeStorageBox>
-            <Text>From total 1Gb</Text>
-          </FreeStorageBox>
-        </FreeStorageContainer>
-        <InputContainer>
-          <SearchItem>
-            <InputItem name="Search" type="text" />
-          </SearchItem>
-          <ButtonSearch>
-            <Button
-              buttoncolor="primary"
-              text="Search"
-              onPress={console.log("pesquisando")}
-            />
-          </ButtonSearch>
-        </InputContainer>
-      </ButtonContainer>
+      <ContainerItems>
+        <UploadButton />
+        <FreeStorage />
+        <SearchComponent id={id} name="Items" />
+      </ContainerItems>
     </Container>
   );
 }