+++ src/screens/SignIn/components/SignInForm/styles.ts
@@ -1,0 +1,32 @@
+import styled from "styled-components/native"
+import { theme } from "../../../../styles/theme"
+
+export const LoginForm = styled.View`
+   align-items: center;
+   width: 100%;
+    
+   `
+export const Inputbox = styled.View`
+    width: 90%;
+    align-items: center;
+    justify-content: center;
+    margin-bottom: 4%;
+`
+export const FourgoutPassBox = styled.View`
+    align-items: flex-end;
+    justify-items: right;
+    width: 90%;
+    margin-bottom: 3%;
+`
+export const ForgoutPass= styled.TouchableOpacity`
+       background-color: none;
+   `
+export const ButtonBox = styled.View`
+    width: 80%;
+    height: 45px;
+    justify-content: center;
+    align-items: center;
+   margin-bottom: 4%;
+   border-radius: 6px;
+`
+