+++ src/screens/SignUp/components/SignUpForm/styles.ts
@@ -1,0 +1,29 @@
+import styled from "styled-components/native"
+
+export const FormSignUp = styled.View`
+    width: 100%;
+    justify-content: center;
+    align-items: center;
+`
+export const InputBox = styled.View`
+    width: 85%;
+    align-items: center;
+    margin-bottom: 4%;
+`
+
+export const BoxText = styled.View`
+    width: 100%;
+    flex-direction: row;
+    justify-content: center;
+`
+export const ButtonItem = styled.View`
+        width: 50%;
+        height: 45px;
+`
+export const TextItem = styled.View`
+    margin-top: 3%;
+    align-items: flex-end;
+`
+export const ButtonSignIn = styled.View`
+    height: 45px;
+`